+++ R/tqDist.r
@@ -29,7 +29,10 @@
 #'   
 #'   \insertRef{Sand2014}{Quartet}
 #' 
-#' @author Martin R. Smith
+#' @seealso [`CompareQuartets`], [`QuartetStatus`]
+#' 
+#' @importFrom ape write.tree
+#' @template MRS
 #' @useDynLib Quartet, .registration = TRUE
 #' @export
 TQDist <- function (trees) {
@@ -141,7 +144,7 @@ SingleTreeQuartetAgreement <- function (trees, comparison = trees[[1]]) {
 #' @templateVar intro Returns a two dimensional array. Rows correspond to the input trees; the first row will report a perfect match if the first tree is specified as the comparison tree (or if `cf` is not specified).  Columns list the status of each quartet:
 #' @template returnEstabrook
 #'         
-#' @author Martin R. Smith
+#' @template MRS
 #' 
 #' @examples
 #'  data('sq_trees')
@@ -154,6 +157,8 @@ SingleTreeQuartetAgreement <- function (trees, comparison = trees[[1]]) {
 #' @family element-by-element comparisons
 #' @seealso `\link{SplitStatus}`: Uses bipartition splits (groups/clades defined by
 #'  nodes or edges of the tree) instead of quartets as the unit of comparison.
+#'  
+#'  [`SimilarityMetrics`]: Generates distance metrics from quartet statuses.
 #' 
 #' @references {
 #'   \insertRef{Brodal2013}{Quartet}
@@ -359,7 +364,7 @@ AllPairsTripletDistance <- function(file) {
 #' @return `TRUE` if `file` is a character vector of length one describing 
 #'   a file that exists, a fatal error otherwise.
 #' 
-#' @author Martin R. Smith
+#' @template MRS
 #' 
 #' @export
 #' @keywords internal