+++ tests/testthat/test-QuartetDistance.R
@@ -30,20 +30,30 @@ test_that("Splits are compared", {
 })
 
 test_that("QuartetStates works", {
-  expect_equal(list(2L), QuartetStates(quartets[[1]]))
-  expect_equal(list(3L), QuartetStates(Tree2Splits(quartets[[2]])))
-  expect_equal(list(2L, 3L, 4L, 0L), QuartetStates(quartets))
+  expect_equal(2L, QuartetStates(quartets[[1]]))
+  expect_equal(3L, QuartetStates(as.Splits(quartets[[2]], letters[1:4])))
+  expect_equal(c(2L, 3L, 4L, 0L), QuartetStates(quartets)[, 1])
 })
 
 test_that("QuartetState works", {
-  expect_equal(2L, QuartetState(letters[1:4], Tree2Splits(quartets[[1]])))
-  expect_equal(3L, QuartetState(letters[1:4], Tree2Splits(quartets[[2]])))
-  expect_equal(4L, QuartetState(letters[1:4], Tree2Splits(quartets[[3]])))
-  expect_equal(0L, QuartetState(letters[1:4], Tree2Splits(quartets[[4]])))
+  expect_equal(2L, QuartetState(letters[1:4], as.Splits(quartets[[1]])))
+  expect_equal(3L, QuartetState(letters[1:4], as.Splits(quartets[[2]])))
+  expect_equal(4L, QuartetState(letters[1:4], as.Splits(quartets[[3]])))
+  expect_equal(0L, QuartetState(letters[1:4], as.Splits(quartets[[4]])))
 })
 
 test_that("CompareQuartets works", {
   expect_equal(c(N=16L, Q=8L, s=1, d=2, r1=1, r2=1, u=3),
                CompareQuartets(c(2, 2, 4, 2, 0, 0, 0, 0),
                                c(2, 3, 3, 0, 2, 0, 0, 0)))
+})
+
+test_that("PlotQuartet works", {
+  library('vdiffr')
+  expect_doppelganger('PlotQuartet', function() {
+    data('sq_trees')
+    
+    par(mfrow=c(3, 2), mar=rep(0.5, 4), cex=1.1)
+    PlotQuartet(sq_trees[c(1, 9, 13:16)], c(2, 5, 3, 8), overwritePar = FALSE)
+  })
 })