+++ NEWS.md
@@ -1,6 +1,7 @@
 # Quartet v1.0.2.9002
 ## Improvements
  - Pass trees directly to C, without writing to temporary intermediate file.
+ - Use `TreeTools` in place of `TreeSearch`, supporting new `Splits` class.
 
 # Quartet v1.0.2.9001
 ## Improvements